requirements.txt
asyncio==3.4.3
openai==0.27.6
pdfplumber==0.9.0
pytesseract==0.3.10
## The following requirements were added by pip freeze:
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==40.0.2
cymem==2.0.7
Deprecated==1.2.13
distlib==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.12.0
frozenlist==1.3.3
googlesearch-python==1.2.3
hiredis==2.2.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.3
opencv-python-headless==4.7.0.72
packaging==23.1
pathy==0.10.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
preshed==3.0.8
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pyjokes==0.6.0
pyproject_hooks==1.0.0
pytest==7.3.1
pytz==2023.3
pywin32-ctypes==0.2.0
redis==4.5.5
redisai==1.3.0
regex==2023.5.5
requests==2.30.0
rfc3986==2.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
tiktoken==0.4.0
tomlkit==0.11.8
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
Wand==0.6.11
wasabi==1.1.1
webencodings==0.5.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0